pyproject.toml
[project]
name = "qr-code-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.127.0",
    "pillow>=12.0.0",
    "python-multipart>=0.0.21",
    "qrcode>=8.2",
    "requests>=2.32.5",
    "types-qrcode>=8.2.0.20250914",
    "types-requests>=2.32.4.20250913",
    "uvicorn>=0.40.0",
]
